Refactor Perf Poetry to add Related Benchmarks
