+++ benchmarks/performance-poetry/complex-poetry/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.squareup.benchmarks.performance.poetry">
+    package="com.squareup.benchmarks.performance.complex.poetry">
 
   <application
       android:allowBackup="false"
@@ -9,7 +9,7 @@
       android:theme="@style/AppTheme"
       tools:ignore="AllowBackup,GoogleAppIndexingWarning,MissingApplicationIcon">
     <activity
-        android:name=".PerformancePoetryActivity"
+        android:name="com.squareup.benchmarks.performance.complex.poetry.PerformancePoetryActivity"
         android:exported="true">
       <intent-filter>
         <action android:name="android.intent.action.MAIN" />
