+++ benchmarks/performance-poetry/complex-poetry/src/main/res/values/strings.xml
